+++ CHANGELOG.md
@@ -8,7 +8,7 @@ Changes since last release
 * Fixed bug when calling `osqp_solve` or `osqp_cleanup` after failed linear system initialization
 * Add "install" CMake target and installation of CMake configuration files
 * Fixed potential name conflict with SCS [47](https://github.com/oxfordcontrol/osqp/issues/47)
-
+* Changed `set_default_settings` to `osqp_set_default_settings` and brought function to main API header `osqp.h`
 
 Version 0.2.1 (25 November 2017)
 ---------------------------------